Added Data Writing to Arduino Control
Added a function to write data to a text file that the script receives
from an arduino. Will go until it times out or receives a command of
STOP from the arduino. Also changed file writing to ‘a’ from ‘w’ so
that it appends rather than writes. (Will this have an issue the first
time when it tries to make the file?)